% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{files_processor}
\alias{files_processor}
\title{Process a single Parquet file}
\usage{
files_processor(file, table_name, rule)
}
\arguments{
\item{file}{Path to one Parquet file.}

\item{table_name}{AEMO table name.}

\item{rule}{A list with \verb{$filter}, \verb{$select}, \verb{$timestamp}.}
}
\value{
A data.table or NULL.
}
\description{
Read entire \code{file}, apply \code{rule}, and return filtered data.table.
For very large tables it delegates to \code{process_file_in_chunks()}.
}
