% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_rules.R
\docType{data}
\name{filtering_rules}
\alias{filtering_rules}
\title{AEMO table‐specific filter/select/timestamp logic}
\format{
A named list of sub‐lists (filter, select, timestamp).
}
\usage{
filtering_rules
}
\description{
A named list (\code{filtering_rules}) where each element corresponds to one AEMO table
(e.g. “BIDDAYOFFER_D”, “DREGION”, etc.).
For each table name, it specifies three functions:
\enumerate{
\item \verb{$filter(dt)}: subset rows.
\item \verb{$select(dt)}: return vector of column names to keep.
\item \verb{$timestamp(dt)}: add or modify a POSIXct \code{timestamp} column.
}
}
\keyword{datasets}
