% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_band_avail_heatmap}
\alias{plot_band_avail_heatmap}
\title{Band Availability Heatmap}
\usage{
plot_band_avail_heatmap(bidper_df, duid_filter)
}
\arguments{
\item{bidper_df}{A data.frame or tibble of BIDPEROFFER_D (must include \code{timestamp} and \verb{BANDAVAIL*} columns)}

\item{duid_filter}{A single DUID string to filter on}
}
\value{
A \strong{ggplot2} object (heatmap)
}
\description{
Plot a heatmap of 5-minute band availability (BANDAVAIL*) over time for one DUID.
}
