% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply_plot.R
\name{plot_supply_curve_animation}
\alias{plot_supply_curve_animation}
\title{Animate Supply Curve}
\usage{
plot_supply_curve_animation(bidday_df, bidper_df, duid_filter)
}
\arguments{
\item{bidday_df}{A data.frame or tibble of BIDDAYOFFER_D}

\item{bidper_df}{A data.frame or tibble of BIDPEROFFER_D}

\item{duid_filter}{Single DUID string}
}
\value{
A \strong{gganimate} object
}
\description{
Create an animated supply curve (cumulative MW vs Price) over time for a single DUID.
}
